### 2021-03-04 #3086 @whedon
[PRE REVIEW]: epipack: An infectious disease modeling package for Python
**Submitting author:** @benmaier (<a href="http://orcid.org/0000-0001-7414-8823">Benjamin F. Maier</a>)
**Repository:** <a href="https://github.com/benmaier/epipack" target ="_blank">https://github.com/benmaier/epipack</a>
**Version:** v0.1.1
