### 2021-03-10 #3097 @chrisvoncsefalvay
The paper does not contain a 'statement of need' section. However, it is my view that the substance of this requirement (pasted below for reference) is addressed adequately in para.s 1 through 3. With that in mind, I would consider requirement 1 of the Documentation section to be met.

> The authors should clearly state what problems the software is designed to solve and who the target audience is.
