### 2021-03-10 #3097 @esteinig
> The authors should clearly state what problems the software is designed to solve and who the target audience is.

Agreed, it is well addressed in the paragraphs
