### 2021-03-18 #3097 @chrisvoncsefalvay
@fboehm I have just completed my review. I've left the installation and functionality boxes last, as I wanted to give a thorough opinion. I have managed to install and test functionalities with success on a Mac OS Big Sur system. I am happy to check off the relevant box based on that, although I would note that the author himself states they have tested it on a Mac OS system – there may be issues on, say, Windows systems that would not have shown up in the course of my testing.

Overall, it has been a pleasure to review what is a well-written, concise yet thorough paper supporting a very useful package that answers a clear need, and does so in a way that bundles often disparate techniques for the analysis of epidemic networks into a single cohesive package.
