### 2021-03-18 3dace20
don't disregard the reset_events flag
