### 2021-03-21 #3097 @esteinig
Hi my apologies for the delay, finishing my thesis at the moment. I am very pleased with the package and the code base, it's already gotten some attention within our lab. I had opportunity to test it on Linux systems but can't make an assessment on other OS.

The paper is well written, and concise. My only recommendation would be to introduce some better structure with headlines in the paper, but that's all :)
