### 2021-03-24 #3097 @fboehm
Thank you, @esteinig! Can you be more specific about the suggestion for headlines in the paper? Can you recommend specific places in paper.md where headlines might be placed?

Thanks again!
