### 2021-03-30 #3097 @benmaier
Dear @fboehm, @chrisvoncsefalvay, and @esteinig, thank you so much for this lightning-fast editing and reviewing work. This has probably been the smoothest reviewing experience I've ever had! I greatly appreciated all your positive comments about the work I submitted.

Regarding adding more headlines, I started working on the manuscript again to incorporate your point, @esteinig, but noticed that I do like the flow of the text as it is. I fear adding more subsections would make this already very short paper even more fragmented. When I wrote it, I had these two successful submissions in mind:

1. https://raw.githubusercontent.com/adrn/gala/f6bb2532afeed3c0b0feea3037bf71a3f42b2a12/paper/paper.md
2. https://github.com/benmaier/netwulf/blob/master/paper/paper.md

where the only subsection was "Summary", as well. I would therefore prefer to keep the text as it is. Of course, this is an editorial decision, ultimately!

(Also, @esteinig, all the best for your thesis!)
