### 2021-03-31 #3097 @fboehm
@benmaier - I think that your solution is reasonable.
