### 2021-04-01 #3097 @fboehm
Great! We now need you, @benmaier to make a new release and to archive the package on zenodo. Please report in this thread the new release's version number and the archive's doi.
