+++ epipack/metadata.py
@@ -3,7 +3,7 @@
 Contains a bunch of information about this package.
 """
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 __author__ = "Benjamin F. Maier"
 __copyright__ = "Copyright 2020, " + __author__
