### 2021-04-01 #3097 @fboehm
I've now downloaded the pdf and verified that all dois in the references section direct to the correct resources.
