### 2021-04-02 #3097 @kthyng
@benmaier Things look like they are mostly ready to go! One comment on your paper. For lists of inline parenthetical references, use this sort of format: [@ref1; @ref2] so that they should up together in one set of parentheses. More info: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
